Change name of LandingPageView to LandingView. Some small styling changes
