--- source/src/main/java/ch/bfh/bti7081/s2020/blue/view/LandingPageView.java
@@ -1,12 +1,0 @@
-package ch.bfh.bti7081.s2020.blue.view;
-
-public interface LandingPageView {
-
-  interface LandingPageViewListener {
-
-  }
-
-
-  public void addListener(LandingPageViewListener listener);
-
-}
