+++ source/src/main/java/ch/bfh/bti7081/s2020/blue/view/MainView.java
@@ -10,7 +10,7 @@ import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.Route;
 import com.vaadin.flow.server.PWA;
 
-@Route
+@Route("home")
 @PWA(name = "Social Anxiety Application",
     shortName = "Social Anxiety App",
     description = "This is an application to help people with social anxiety.",
